@@ -11,20 +11,20 @@
 		"publish": "vite build && gh-pages -d build -t true"
 	},
 	"devDependencies": {
-		"@fontsource/roboto-condensed": "^5.0.5",
-		"@rollup/plugin-dsv": "^3.0.2",
-		"@sveltejs/adapter-static": "^2.0.2",
-		"@sveltejs/kit": "^1.5.0",
+		"@fontsource/roboto-condensed": "^5.1.1",
+		"@rollup/plugin-dsv": "^3.0.5",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/kit": "^2.17.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"bulma": "^0.9.4",
 		"d3-array": "^3.2.4",
 		"d3-dsv": "^3.0.1",
-		"gh-pages": "^5.0.0",
+		"gh-pages": "^6.3.0",
 		"html-minifier-terser": "^7.2.0",
-		"prettier": "^3.0.0",
-		"prettier-plugin-svelte": "^3.0.3",
-		"sass": "^1.63.6",
-		"svelte": "^4.1.1",
-		"vite": "^4.3.0",
-		"vite-imagetools": "^5.0.7"
+		"prettier": "^3.5.0",
+		"prettier-plugin-svelte": "^3.3.3",
+		"sass": "^1.84.0",
+		"svelte": "^5.20.0",
+		"vite-imagetools": "^7.0.5"
 	}
 }
