@@ -1,23 +1,28 @@
-<script>
+<script lang="ts">
+	import { run } from 'svelte/legacy';
+
 	import { base } from '$app/paths';
 	import { page } from '$app/stores';
 
-	export let classes;
-	export let href;
-	export let target;
+	let {
+		classes,
+		href = $bindable(),
+		target = $bindable(),
+		children
+	} = $props();
 
-	let active;
-	let external;
-	$: {
+	let active = $state();
+	let external = $state();
+	run(() => {
 		external = href.startsWith('http');
 		target = external ? '_blank' : null;
 		href = external ? href : base + href;
 		active = $page.url.pathname === href;
-	}
+	});
 </script>
 
 <a {href} aria-current={active} class="navbar-item {classes}" class:is-active={active} {target}
-	><slot /></a
+	>{@render children?.()}</a
 >
 
 <style>
