@@ -4,13 +4,5 @@ import { imagetools } from 'vite-imagetools';
 import dsv from '@rollup/plugin-dsv';
 
 export default defineConfig({
-	plugins: [sveltekit(), imagetools(), dsv()],
-
-	css: {
-		preprocessorOptions: {
-			scss: {
-				additionalData: '@use "src/variables.scss" as *;'
-			}
-		}
-	}
+	plugins: [sveltekit(), imagetools(), dsv()]
 });
